Add the execd.py module to wazuh_testing and move the auxiliary functions from the conftest.py of test_execd to it #1071
